Generate neutral and primary scales from Radix and output as css vars
