+++ packages/theme/src/index.tsx
@@ -1,5 +1,6 @@
 import { forwardRef } from 'react';
 import { useGenerateStyles } from './color/a8c';
+import { useGenerateRadixStyles } from './color/radix';
 import type { ThemeProps } from './types';
 
 declare module 'react' {
@@ -10,9 +11,10 @@ declare module 'react' {
 
 const Theme = forwardRef< HTMLDivElement, ThemeProps >( function Theme( { color, children }, ref ) {
 	const styles = useGenerateStyles( color );
+	const radixColors = useGenerateRadixStyles( color );
 
 	return (
-		<div ref={ ref } style={ styles }>
+		<div ref={ ref } style={ { ...styles, ...radixColors } }>
 			{ children }
 		</div>
 	);
